+++ src/Diagnostic/DiagnosticParticleBinningBase.h
@@ -8,6 +8,7 @@
 class DiagnosticParticleBinningBase : public Diagnostic
 {
     friend class SmileiMPI;
+    friend class VectorPatch;
     
 public :
 
@@ -31,6 +32,8 @@ public :
     
     bool prepare( int itime ) override;
     
+    void calculate_auto_limits( Patch *patch, SimWindow *simWindow, unsigned int ipatch );
+    
     virtual void run( Patch *patch, int itime, SimWindow *simWindow ) override;
     
     virtual bool writeNow( int itime );
@@ -51,6 +54,8 @@ public :
     
     //! Get disk footprint of current diagnostic
     uint64_t getDiskFootPrint( int istart, int istop, Patch *patch ) override;
+    
+    std::vector<std::vector<double> > patches_mins, patches_maxs;
     
 protected: