+++ src/Diagnostic/HistogramFactory.h
@@ -109,7 +109,7 @@ public:
             
             histogram->axes.push_back( axis );
         }
-
+        
         return histogram;
     }
     
@@ -124,7 +124,7 @@ public:
     )
     {
         HistogramAxis *axis = nullptr;
-        std::string type = "";
+        std::string type = "", automin, automax;
         double min, max;
         int nbins;
         bool logscale, edge_inclusive;
@@ -168,26 +168,34 @@ public:
                 ParticleData test( params.nDim_particle, type_object, typePrefix.str(), dummy );
                 type ="user_function";
 #else
-                ERROR( errorPrefix << ": First item must be a string (axis type)" );
+                ERROR( errorPrefix << ": First item (axis type) must be a string" );
 #endif
             }
         }
         
         // Try to extract second element: axis min
         if( !PyTools::py2scalar( PySequence_Fast_GET_ITEM( seq, i ), min ) ) {
-            ERROR( errorPrefix<< ": Second item must be a double (axis min)" );
+            if( PyTools::py2scalar( PySequence_Fast_GET_ITEM( seq, i ), automin ) && automin == "auto" ) {
+                min = nan("");
+            } else {
+                ERROR( errorPrefix<< ": Second item (axis min) must be a float or 'auto'" );
+            }
         }
         i++;
         
         // Try to extract third element: axis max
         if( !PyTools::py2scalar( PySequence_Fast_GET_ITEM( seq, i ), max ) ) {
-            ERROR( errorPrefix << ": Third item must be a double (axis max)" );
+            if( PyTools::py2scalar( PySequence_Fast_GET_ITEM( seq, i ), automax ) && automax == "auto" ) {
+                max = nan("");
+            } else {
+                ERROR( errorPrefix << ": Third item (axis max) must be a float or 'auto'" );
+            }
         }
         i++;
         
         // Try to extract fourth element: axis nbins
         if( !PyTools::py2scalar( PySequence_Fast_GET_ITEM( seq, i ), nbins ) ) {
-            ERROR( errorPrefix << ": Fourth item must be an int (number of bins)" );
+            ERROR( errorPrefix << ": Fourth item (number of bins) must be an integer" );
         }
         i++;