+++ doc/Sphinx/releases.rst
@@ -23,6 +23,7 @@ Changes made in the repository (not released)
 * Lasers can be injected from all boundaries
 * Probes can include components of the Poynting vector ``PoyX``, ``PoyY``, ``PoyZ``
 * Probes can be time-integrated
+* ``ParticleBinning`` diagnostics may accept ``"auto"`` as axis limits
 * Bugfixes:
 
   * Poynting scalars with checkpoints